v1.1.0

+++ app/src/main/java/com/cabovianco/remindme/data/alarm/AlarmScheduler.kt
@@ -23,7 +23,7 @@ class AlarmScheduler @Inject constructor(
 
         val pendingIntent = PendingIntent.getBroadcast(
             context,
-            id,
+            id.toInt(),
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
@@ -41,12 +41,12 @@ class AlarmScheduler @Inject constructor(
         )
     }
 
-    fun cancel(id: Int) {
+    fun cancel(id: Long) {
         val intent = Intent(context, NotificationReceiver::class.java)
 
         val pendingIntent = PendingIntent.getBroadcast(
             context,
-            id,
+            id.toInt(),
             intent,
             PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
         )
